working demo

+++ CodeThe_Rope/goal.pde
@@ -1,17 +1,16 @@
-public class goal extends node{
+public class goal extends node {
   float radius;
   PImage img;
-  public goal (float x, float y, float radius_){
-    super(x,y,createShape(ELLIPSE, 0, 0, radius_, radius_),color(0,255,0));
+  public goal (float x, float y, float radius_) {
+    super(x, y, createShape(ELLIPSE, 0, 0, radius_, radius_), color(0, 255, 0));
     radius=radius_;
     img = loadImage("omNom.png");
   }
-  void display(){
+  void display() {
     imageMode(CENTER);
-    image(img, position.x, position.y,radius+10,radius+10);
+    image(img, (float) position.x, (float) position.y, radius+10, radius+10);
   }
-  public float getRadius(){
+  public float getRadius() {
     return radius;
   }
-  
 }
